+++ Tree_test.go
@@ -1,29 +1,10 @@
-package Node
+package XMLHandler
 
 import (
 	"fmt"
 )
 
-func iAddAChildNodeLabelledToTheNodeTree(arg1 string) error {
-
-	if nodeTree == nil {
-		return fmt.Errorf("no node tree")
-	}
-
-	if shouldFail == true {
-		shouldFail = false
-		return fmt.Errorf("this step does not know how to fail")
-	}
-
-	newNode := Node{}
-	if e := nodeTree.AddChild(&newNode); e != nil {
-		return e
-	}
-	newNode.label = arg1
-	return nil
-
-}
-
+//base
 func nodeShouldHaveTheParent(arg1, arg2 string) error {
 
 	if nodeTree == nil {