+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## [0.2.0]
+
+### Added:
+- Added functions `ConvertNodeToXML`, `ConvertTreeToXML`, `ExportTree`, `ImportFile`, `SortXML`
+- Removed redundant package `Node` and moved its contents to base package `XMLHandler`
+- Cleaned up `_test.go` files
+
 ## [0.1.4] - 2020/04/03
 
 ### Added: